// src/models/enums/ErrorCode.ts
export enum ErrorCode {
    invalidCredentials = "INVALID_CREDENTIALS",
    tokenMissing = "TOKEN_MISSING",
    tokenInvalid = "TOKEN_INVALID",
    serverError = "SERVER_ERROR",
    payloadUndefined = "PAYLOAD_UNDEFINED",
    moduleNotFound = "MODULE_NOT_FOUND",
    invalidModuleId = "INVALID_MODULE_ID",
    invalidUser = "INVALID_USER",
    emailNotVerified = "EMAIL_NOT_VERIFIED",
    recaptchaFailed = "RECAPTCHA_FAILED",
    invalidContents = "INVALID_CONTENTS",
    invalidModuleData = "INVALID_MODULE_DATA",
    invalidContentData = "INVALID_CONTENT_DATA",
    invalidContentItem = "INVALID_CONTENT_ITEM",
    unknownError = "UNKNOWN_ERROR",
    invalidContentId = "INVALID_CONTENT_ID",
    contentNotFound = "CONTENT_NOT_FOUND",
    unauthorizedUser = "USER_UNAUTHORIZED",
    invalidUserId = "INVALID_USER_ID",
    progressNotFound = "PROGRESS_NOT_FOUND",
    invalidInputData = "INVALID_INPUT_DATA"
}